refactor: consolidate Mailchimp list types into single file
- Merge list.ts and lists-page-props.ts into lists.ts for better organization
- Update all imports to use consolidated lists.ts module
- Remove deprecated ListsParamsInternalSchema from schema file
- Update service exports to use ListsParams instead of ListsQuery
- Maintain consistent naming with ListsParams throughout codebase
