+++ src/app/mailchimp/lists/page.test.tsx
@@ -38,8 +38,8 @@ vi.mock("@/components/dashboard/shared/dashboard-error", () => ({
 // Import the components we want to test
 import { ListStats } from "@/components/mailchimp/lists/ListStats";
 import { ClientListList } from "@/components/mailchimp/lists/ClientListList";
-import type { List } from "@/types/mailchimp/list";
-import type { DashboardListStats as ListStatsType } from "@/types/mailchimp/list";
+import type { List } from "@/types/mailchimp/lists";
+import type { DashboardListStats as ListStatsType } from "@/types/mailchimp/lists";
 
 const mockLists: List[] = [
   {
